requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
binance-futures-connector==3.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
constantly==15.1.0
cryptography==41.0.3
dateparser==1.1.8
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
frozenlist==1.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.7
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pandas-ta==0.3.14b0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.16.1
pyOpenSSL==23.2.0
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.1.1
regex==2023.6.3
requests==2.31.0
service-identity==23.1.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.4
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.2
zope.interface==6.0
